requirements.txt
pytesseract
pdf2image
Pillow
PyMuPDF
pdfplumber
opencv-python
numpy
easyocr
